Разбил содержимое CircularCloudLayouter.cs на отдельные файлы
