+++ cs/TagsCloudVisualization/Directions.cs
@@ -1,0 +1,10 @@
+﻿namespace TagsCloudVisualization
+{
+    public enum Directions
+    {
+        Up,
+        Left,
+        Down,
+        Right
+    }
+}
