+++ cs/TagsCloudVisualization/TagsCloudVisualization.csproj
@@ -60,7 +60,11 @@
     </ItemGroup>
     <ItemGroup>
         <Compile Include="CircularCloudLayouter.cs" />
+        <Compile Include="CircularCloudLayouter_Should.cs" />
+        <Compile Include="Directions.cs" />
+        <Compile Include="DirectionsExtensions.cs" />
         <Compile Include="Properties\AssemblyInfo.cs" />
+        <Compile Include="RectangleExtensions.cs" />
     </ItemGroup>
     <ItemGroup>
       <None Include="packages.config" />
